requirements.txt
Django>=5.2.2
djangorestframework>=3.15.1
drf-yasg>=1.21.7 